@@ -1,9 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-		<TargetFrameworks>net6.0;net7.0;netstandard2.1</TargetFrameworks>
         <ImplicitUsings>disable</ImplicitUsings>
-        <LangVersion>8.0</LangVersion>
+        <LangVersion>13</LangVersion>
         <RootNamespace>TonSdk.Client</RootNamespace>
         <PackageId>TonSdk.Client</PackageId>
         <Version>0.3.11</Version>
@@ -11,12 +10,13 @@
         <PackageDescription>Ton Client library for work with HTTP API and Lite Servers.</PackageDescription>
         <RepositoryUrl>https://github.com/continuation-team/TonSdk.NET</RepositoryUrl>
         <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+        <TargetFramework>net9.0</TargetFramework>
+        <Nullable>enable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
         <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-        <PackageReference Include="TonSdk.Adnl" Version="0.2.4" />
-        <PackageReference Include="TonSdk.Core" Version="0.3.9" />
+        <ProjectReference Include="../../TonSdk.Adnl/src/TonSdk.Adnl.csproj " />
     </ItemGroup>
 
     <PropertyGroup>
