@@ -1,91 +1,90 @@
 ﻿using System;
 using System.Security.Cryptography;
 
-namespace TonSdk.Adnl
+namespace TonSdk.Adnl.Adnl;
+
+public class AesCounter
 {
-
-    public class AesCounter
+    public AesCounter(byte[] initialValue)
     {
-        private readonly byte[] _counter;
-
-        public AesCounter(byte[] initialValue)
-        {
-            if (initialValue.Length != 16)
-                throw new ArgumentException("Invalid counter bytes size (must be 16 bytes)");
-            _counter = initialValue;
-        }
-
-        public AesCounter(int initialValue)
-        {
-            _counter = new byte[16];
-            for (int i = 15; i >= 0; i--)
-            {
-                _counter[i] = (byte)(initialValue % 256);
-                initialValue /= 256;
-            }
-        }
-
-        public void Increment()
-        {
-            for (int i = 15; i >= 0; i--)
-            {
-                if (_counter[i] == 255)
-                    _counter[i] = 0;
-                else
-                {
-                    _counter[i]++;
-                    break;
-                }
-            }
-        }
-
-        public byte[] Counter => _counter;
+        if (initialValue.Length != 16)
+            throw new ArgumentException("Invalid counter bytes size (must be 16 bytes)");
+        Counter = initialValue;
     }
 
-    public class AesCtrMode
+    public AesCounter(int initialValue)
     {
-        private AesCounter _counter;
-        private byte[] _remainingCounter;
-        private int _remainingCounterIndex;
-        private Aes _aes;
-
-        public AesCtrMode(byte[] key, AesCounter? counter)
+        Counter = new byte[16];
+        for (var i = 15; i >= 0; i--)
         {
-            _counter = counter ?? new AesCounter(1);
-            _remainingCounter = new byte[16];
-            _remainingCounterIndex = 16;
-
-            _aes = Aes.Create();
-            _aes.Key = key;
-            _aes.Mode = CipherMode.ECB;
-            _aes.Padding = PaddingMode.None;
+            Counter[i] = (byte)(initialValue % 256);
+            initialValue /= 256;
         }
+    }
 
-        public byte[] Encrypt(byte[] plaintext)
-        {
-            byte[] encrypted = new byte[plaintext.Length];
+    public byte[] Counter { get; }
 
-            for (int i = 0; i < encrypted.Length; i++)
+    public void Increment()
+    {
+        for (var i = 15; i >= 0; i--)
+            if (Counter[i] == 255)
             {
-                if (_remainingCounterIndex == 16)
-                {
-                    _remainingCounter = EncryptCounter(_counter.Counter);
-                    _remainingCounterIndex = 0;
-                    _counter.Increment();
-                }
+                Counter[i] = 0;
+            }
+            else
+            {
+                Counter[i]++;
+                break;
+            }
+    }
+}
 
-                encrypted[i] = (byte)(plaintext[i] ^ _remainingCounter[_remainingCounterIndex++]);
+public class AesCtrMode
+{
+    private readonly Aes _aes;
+    private readonly AesCounter _counter;
+    private byte[] _remainingCounter;
+    private int _remainingCounterIndex;
+
+    public AesCtrMode(byte[] key, AesCounter? counter)
+    {
+        _counter = counter ?? new AesCounter(1);
+        _remainingCounter = new byte[16];
+        _remainingCounterIndex = 16;
+
+        _aes = Aes.Create();
+        _aes.Key = key;
+        _aes.Mode = CipherMode.ECB;
+        _aes.Padding = PaddingMode.None;
+    }
+
+    public byte[] Encrypt(byte[] plaintext)
+    {
+        var encrypted = new byte[plaintext.Length];
+
+        for (var i = 0; i < encrypted.Length; i++)
+        {
+            if (_remainingCounterIndex == 16)
+            {
+                _remainingCounter = EncryptCounter(_counter.Counter);
+                _remainingCounterIndex = 0;
+                _counter.Increment();
             }
 
-            return encrypted;
+            encrypted[i] = (byte)(plaintext[i] ^ _remainingCounter[_remainingCounterIndex++]);
         }
 
-        private byte[] EncryptCounter(byte[] counter)
-        {
-            using var encryptor = _aes.CreateEncryptor();
-            return encryptor.TransformFinalBlock(counter, 0, counter.Length);
-        }
+        return encrypted;
+    }
 
-        public byte[] Decrypt(byte[] ciphertext) => Encrypt(ciphertext);
+    private byte[] EncryptCounter(byte[] counter)
+    {
+        using var encryptor = _aes.CreateEncryptor();
+        return encryptor.TransformFinalBlock(counter, 0, counter.Length);
+    }
+
+    public byte[] Decrypt(byte[] ciphertext)
+    {
+        return Encrypt(ciphertext);
     }
 }
