@@ -1,7 +1,9 @@
 using NUnit.Framework;
+using TonSdk.Adnl.Adnl;
 using TonSdk.Core.Crypto;
 
 namespace TonSdk.Adnl.Tests;
+
 public class AdnlTest
 {
     private readonly string AdnlHost = "5.9.10.47";
@@ -15,22 +17,21 @@ public class AdnlTest
     {
         client = new AdnlClientTcp(AdnlHost, AdnlPort, AdnlPubKey);
     }
-    
+
     [Test]
     public async Task Test_ConnectionState()
     {
         Assert.That(AdnlClientState.Closed == client.State, Is.EqualTo(true));
         await client.Connect();
         Assert.That(AdnlClientState.Connecting == client.State, Is.EqualTo(true));
-        
+
         while (client.State != AdnlClientState.Open)
         {
             // waiting
-            continue;
         }
-        
+
         Assert.That(AdnlClientState.Open == client.State, Is.EqualTo(true));
-        
+
         client.End();
         Assert.That(AdnlClientState.Closed == client.State, Is.EqualTo(true));
     }
@@ -39,23 +40,22 @@ public class AdnlTest
     public async Task Test_GetDataTest()
     {
         byte[]? data = null;
-        client.DataReceived += (response) => data = response;
+        client.DataReceived += response => data = response;
         await client.Connect();
-        
+
         while (client.State != AdnlClientState.Open)
         {
             // waiting
-            continue;
         }
 
-        await client.Write(Utils.HexToBytes(
+        await client.Write(TonSdk.Core.Crypto.Utils.HexToBytes(
             "7af98bb435263e6c95d6fecb497dfd0aa5f031e7d412986b5ce720496db512052e8f2d100cdf068c7904345aad16000000000000"));
-        
+
         while (data == null)
         {
             // waiting
-            continue;
         }
+
         Assert.That(data.Length != 0, Is.EqualTo(true));
     }
 }
